Struct osr_extract::Replay

pub(crate) struct Replay {

    pub(crate) game_mode: u8,
    pub(crate) version: u32,
    pub(crate) beatmap_hash: String,
    pub(crate) player_name: String,
    pub(crate) replay_hash: String,
    pub(crate) count300: u16,
    pub(crate) count100: u16,
    pub(crate) count50: u16,
    pub(crate) count_geki: u16,
    pub(crate) count_katu: u16,
    pub(crate) count_miss: u16,
    pub(crate) score: u32,
    pub(crate) max_combo: u16,
    pub(crate) perfect_combo: bool,
    pub(crate) mods: u32,
    pub(crate) hp: String,
    pub(crate) timestamp: u64,
    pub(crate) replay_lenth: u32,
    pub(crate) replay_data: Vec<u8>,

}

Fields
game_mode: u8

Game mode of the replay (0 = osu! Standard, 1 = Taiko, 2 = Catch the Beat, 3 = osu!mania)

version: u32

Version of the game when the replay was created (ex. 20131216)

beatmap_hash: String

osu! beatmap MD5 hash

player_name: String

Player name

replay_hash: String

osu! replay MD5 hash (includes certain properties of the replay)

count300: u16

Number of 300s

count100: u16

Number of 100s in standard, 150s in Taiko, 100s in CTB, 100s in mania

count50: u16

Number of 50s in standard, small fruit in CTB, 50s in mania

count_geki: u16

Number of Gekis in standard, Max 300s in mania

count_katu: u16

Number of Katus in standard, 200s in mania

count_miss: u16

Number of misses

score: u32

Total score displayed on the score report

max_combo: u16

Greatest combo displayed on the score report

perfect_combo: bool

Perfect/full combo (1 = no misses and no slider breaks and no early finished sliders)

mods: u32

Mods used. See below for list of mod values.

hp: String

Life bar graph: comma separated pairs u/v, where u is the time in milliseconds into the song and v is a floating point value from 0 - 1 that represents the amount of life you have at the given time (0 = life bar is empty, 1= life bar is full)

timestamp: u64

Time stamp (Windows ticks)

replay_lenth: u32

Length in bytes of compressed replay data

replay_data: Vec<u8>

Compressed replay data